requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.9
openai>=1.30.0
python-docx==1.1.2
PyPDF2==3.0.1
reportlab==4.2.5
python-dotenv==1.0.1
openpyxl==3.1.5
